Fix perm loading
